Move VAOS unit tests: Batch 1
